+++ src/config.ts
@@ -40,7 +40,6 @@ const LOCALE_REGEX = /^[a-z]{2}(-[A-Za-z]{2})?$/;
 const LOG_PREFIX = '[Config]';
 
 type ApiRequestDefaults = {
-    allowBackgroundLoad?: boolean;      // If true, this endpoint will return the cache immediately and then load more data in the background
     initialMaxRecords?: number;         // How many records to pull in the initial bootstrap
     limit?: number;                     // Max per page is usually 40
     lookbackForUpdatesMinutes?: number; // How long to look back for updates (edits, increased reblogs, etc.)
@@ -173,7 +172,8 @@ interface ConfigType {
  */
 class Config implements ConfigType {
     api = {
-        backgroundLoadSleepBetweenRequestsMS: 1_000, // How long to wait between API requests during backgrund load
+        // How long to wait between API requests during backgrund load (actually a random number between 0 and this value)
+        backgroundLoadSleepBetweenRequestsMS: 1_200,
         backgroundLoadIntervalMinutes: 10,      // Time between background polling for additional user data after initial load
         daysBeforeFullCacheRefresh: 21,         // Days before the cache is considered stale and needs to be refreshed completely  // TODO: currently unused
         defaultRecordsPerPage: 40,              // Max per page is usually 40: https://docs.joinmastodon.org/methods/timelines/#request-2
@@ -202,7 +202,6 @@ class Config implements ConfigType {
                 minutesUntilStale: 12 * MINUTES_IN_HOUR,
             },
             [CacheKey.FOLLOWED_ACCOUNTS]: {
-                allowBackgroundLoad: true,
                 initialMaxRecords: 1_600,
                 limit: 80,
                 minutesUntilStale: 12 * MINUTES_IN_HOUR,